\name{Math-methods}
\docType{methods}
\alias{Math-methods}
\alias{Math,vf-method}
\alias{Arith-methods}
\alias{Arith,function,vf-method}
\alias{Arith,ANY,vf-method}
\alias{Arith,vf,function-method}
\alias{Arith,vf,missing-method}
\alias{Arith,vf,ANY-method}
\alias{Arith,vf,vf-method}
\alias{Compare-methods}
\alias{Compare,function,vf-method}
\alias{Compare,ANY,vf-method}
\alias{Compare,vf,function-method}
\alias{Compare,vf,ANY-method}
\alias{Compare,vf,vf-method}
\title{Methods for Function \code{Math}, \code{Compare}, \code{Arith} in
  the \code{vfunc} package}
\description{
Various \proglang{S4} methods to work with \code{vf} objects
}
\keyword{methods}

